fix glm_ivec2|3_fill and glm_ivec2|3_eq params continue

+++ include/cglm/call/ivec2.h
@@ -63,11 +63,11 @@ glmc_ivec2_distance(ivec2 a, ivec2 b);
 
 CGLM_EXPORT
 void
-glmc_ivec2_fill(ivec2 v, float val);
+glmc_ivec2_fill(ivec2 v, int val);
 
 CGLM_EXPORT
 bool
-glmc_ivec2_eq(ivec2 v, float val);
+glmc_ivec2_eq(ivec2 v, int val);
 
 CGLM_EXPORT
 bool
